Add setup for oh-my-bash + format + powerlevel10k update

+++ uninstall.sh
@@ -1,13 +1,45 @@
 #!/bin/sh
 
-# remove settings files
-rm ~/.zshrc
-rm ~/.p10k*
-rm ~/.zsh-autosuggestions.zsh
+# Uninstall zsh setup
+if [ -n $(which zsh) ]; then
+    echo "Unnstalling zsh setup"
+    # remove settings files
+    rm ~/.zshrc
+    rm ~/.p10k*
+    rm ~/.zsh-autosuggestions.zsh
+    # uninstall oh-my-zsh
+    ZSH=~/.setup/.oh-my-zsh sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/uninstall.sh)"
+    rm -rf .oh-my-zsh
+    # If original exists, restore
+    if [ -f .zshrc.orig ]; then
+        cp .zshrc.orig ~/.zshrc
+    fi
+fi
+
+# Uninstall bash setup
+if [ -n $(which bash) ]; then
+    echo "Unnstalling bash setup"
+    # remove settings files
+    rm ~/.bashrc
+    # uninstall oh-my-bash
+    OSH="$HOME/.setup/.oh-my-bash" sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmybash/oh-my-bash/master/tools/uninstall.sh)"
+    rm -rf .oh-my-bash
+    # If original exists, restore
+    if [ -f .bashrc.orig ]; then
+        cp .bashrc.orig ~/.bashrc
+    fi
+fi
+
+# Uninstall vim setup
+if [ -n $(which vim) ]; then
+    echo "Unnstalling vim setup"
+    # remove settings files
+    rm ~/.vimrc
+    # If original exists, restore
+    if [ -f .vimrc.orig ]; then
+        cp .vimrc.orig ~/.vimrc
+    fi
+fi
 
 # uninstall fzf
 ./third-party/fzf/uninstall
-
-# uninstall oh-my-zsh
-ZSH=~/.setup/.oh-my-zsh sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/uninstall.sh)"
-rm -rf .oh-my-zsh
